task: moved filtering logic to utils
